Fix type conversion failure in TBufferFile::ReadObjectAny.
TBufferFile::ReadObjectAny does not correctly support schema evolution,
because the fact that there is a conversion never gets passed to the
streamer.

This was originally reported in ROOT-8367.  See there for a reproducer.
This implements the fix proposed in that report.

ATLAS trigger code relies on this.  I currently have a workaround in ATLAS,
but i'd like to be able to remove that.

Should probably go into 6.24-patches.